Reduced Code by 352 lines: Extract and centralize common logic related to creating actions, generating F-Curves for bones, and inserting keyframes into a dedicated module. This makes it easier to modify animation behavior and simplifies the main import functions.
